pyproject.toml
[project]
name = "Llamora"
version = "0.1.0"
description = "A chatbot webapp based on Python, HTMX, and a llamafile server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.4",
    "python-dotenv>=1.1.1",
    "pynacl>=1.5.0",
    "aiosqlite>=0.20.0",
    "python-ulid>=3.0.0",
    "quart>=0.20.0",
    "quart-wtforms>=1.0.3",
    "httpx>=0.27.0",
    "zxcvbn>=4.4.2",
    "aiosqlitepool>=1.0.0",
    "numpy>=1.26.4",
    "hnswlib>=0.8.0",
    "pyahocorasick>=2.1.0",
    "fastembed>=0.7.1",
    "humanize>=4.9.0",
    "cachetools>=5.3.3",
    "orjson>=3.10.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pyright>=1.1.407",
    "pytest>=8.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src/llamora"]

[tool.hatch.build.targets.wheel]
packages = ["src/llamora"]

[tool.hatch.build.targets.sdist]
include = ["src/llamora"]
